Seperated dynamic fields rendering logic into dedicated components
