+++ packages/commonwealth/client/scripts/views/pages/CreateQuest/QuestForm/QuestActionSubForm/SpecialCaseDynamicFields/types.ts
@@ -1,0 +1,6 @@
+import { QuestActionSubFormProps } from '../types';
+
+export type SpecialCaseDynamicFieldsProps = Pick<
+  QuestActionSubFormProps,
+  'defaultValues' | 'errors' | 'onChange' | 'config'
+>;
