+++ packages/commonwealth/client/scripts/views/pages/CreateQuest/QuestForm/QuestActionSubForm/useQuestActionSubForm.ts
@@ -2,19 +2,14 @@ import {
   QuestParticipationLimit,
   QuestParticipationPeriod,
 } from '@hicommonwealth/schemas';
-import { PRODUCTION_DOMAIN } from '@hicommonwealth/shared';
 import { numberNonDecimalGTZeroValidationSchema } from 'helpers/formValidations/common';
 import { splitCamelOrPascalCase } from 'helpers/string';
 import useRunOnceOnCondition from 'hooks/useRunOnceOnCondition';
 import { useEffect } from 'react';
-import { fetchCachedNodes } from 'state/api/nodes';
 import { useCWRepetitionCycleRadioButton } from 'views/components/component_kit/CWRepetitionCycleRadioButton';
 import { ValidationFnProps } from 'views/components/component_kit/CWRepetitionCycleRadioButton/types';
-import {
-  QuestAction,
-  QuestActionContentIdScope,
-  QuestActionSubFormProps,
-} from './types';
+import { doesConfigAllowContentIdField } from '../helpers';
+import { QuestAction, QuestActionSubFormProps } from './types';
 
 // these restrictions are only on client side, update per future requirements
 const MAX_REPETITION_COUNTS = {
@@ -38,120 +33,9 @@ const useQuestActionSubForm = ({
     }))
     .filter((action) => !(hiddenActions || []).includes(action.value));
 
-  const inputConfigs = {
-    contentId: {
-      placeholders: {
-        sampleThreadLink: `https://${PRODUCTION_DOMAIN}/discussion/25730`,
-        sampleCommentLink: `https://${PRODUCTION_DOMAIN}/discussion/25730?comment=89775`,
-        sampleTopicLink: `https://${PRODUCTION_DOMAIN}/common/discussions/Proposals`,
-        twitterTweetUrl: `https://x.com/user/status/1904060455158428146`,
-        discordServerId: `0xxxxxxxxxxxxxxxx0`,
-        chainId: `Select community chain`,
-        groupId: `https://${PRODUCTION_DOMAIN}/common/members?tab=groups&groupId=1234`,
-      },
-      labels: {
-        threadId: 'Thread Link (optional)',
-        commentId: 'Comment Link (optional)',
-        topicId: 'Topic Link (optional)',
-        twitterTweetUrl: 'Tweet URL',
-        chainId: 'Chain Id (optional)',
-        discordServerId: 'Discord Server Id',
-        groupId: 'Group Link',
-      },
-    },
-    startLink: {
-      placeholders: {
-        discordServerUrl: `https://discord.gg/commonwealth`,
-      },
-      labels: {
-        discordServerUrl: 'Discord Server Url',
-      },
-    },
-  };
-
-  const getContentIdInputConfig = () => {
-    if (defaultValues?.contentIdScope === QuestActionContentIdScope.Thread) {
-      if (config?.with_optional_thread_id) {
-        return {
-          label: inputConfigs.contentId.labels.threadId,
-          placeholder: inputConfigs.contentId.placeholders.sampleThreadLink,
-        };
-      }
-      if (config?.with_optional_comment_id) {
-        return {
-          label: inputConfigs.contentId.labels.commentId,
-          placeholder: inputConfigs.contentId.placeholders.sampleCommentLink,
-        };
-      }
-    }
-    if (
-      config?.with_optional_topic_id ||
-      defaultValues?.contentIdScope === QuestActionContentIdScope.Topic
-    ) {
-      return {
-        label: inputConfigs.contentId.labels.topicId,
-        placeholder: inputConfigs.contentId.placeholders.sampleTopicLink,
-      };
-    }
-
-    if (config?.requires_twitter_tweet_link) {
-      return {
-        label: inputConfigs.contentId.labels.twitterTweetUrl,
-        placeholder: inputConfigs.contentId.placeholders.twitterTweetUrl,
-      };
-    }
-
-    if (config?.requires_discord_server_id) {
-      return {
-        label: inputConfigs.contentId.labels.discordServerId,
-        placeholder: inputConfigs.contentId.placeholders.discordServerId,
-      };
-    }
-
-    if (config?.with_optional_chain_id) {
-      return {
-        label: inputConfigs.contentId.labels.chainId,
-        placeholder: inputConfigs.contentId.placeholders.chainId,
-      };
-    }
-
-    if (config?.requires_group_id) {
-      return {
-        label: inputConfigs.contentId.labels.groupId,
-        placeholder: inputConfigs.contentId.placeholders.groupId,
-      };
-    }
-
-    return { label: 'Content Id', placeholder: 'Content Id' };
-  };
-
-  const getStartLinkInputConfig = () => {
-    if (config?.requires_discord_server_id) {
-      return {
-        label: inputConfigs.startLink.labels.discordServerUrl,
-        placeholder: inputConfigs.startLink.placeholders.discordServerUrl,
-      };
-    }
-
-    return { label: 'Start Link', placeholder: 'https://example.com' };
-  };
-
-  const hasContentIdField =
-    config?.with_optional_comment_id ||
-    config?.with_optional_thread_id ||
-    config?.with_optional_topic_id ||
-    config?.requires_twitter_tweet_link ||
-    config?.requires_discord_server_id ||
-    config?.with_optional_chain_id ||
-    config?.requires_group_id;
-
-  const communityChainNodeSelectInputOptions = fetchCachedNodes()
-    ?.filter((node) => node?.ethChainId || node?.cosmosChainId)
-    ?.map((node) => ({
-      label: `${node.name} - (Chain Id = ${node?.ethChainId || node?.cosmosChainId})`,
-      value: node.id,
-    }))
-    ?.sort((a, b) => a.label.localeCompare(b.label));
+  const hasContentIdField = config
+    ? doesConfigAllowContentIdField(config)
+    : false;
 
   const repetitionCycleOptions = Object.keys(QuestParticipationPeriod).map(
     (k) => ({
@@ -300,12 +184,7 @@ const useQuestActionSubForm = ({
     doesActionPreventRepetition,
     repetitionCycleRadio,
     actionOptions,
-    inputConfigs: {
-      contentId: getContentIdInputConfig(),
-      startLink: getStartLinkInputConfig(),
-    },
     hasContentIdField,
-    communityChainNodeSelectInputOptions,
   };
 };
